refactor: Auth와 Member 기능 분리
