+++ src/main/java/woowacourse/auth/application/AuthService.java
@@ -5,17 +5,11 @@ import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.dao.MemberDao;
 import woowacourse.auth.domain.Email;
 import woowacourse.auth.domain.Member;
-import woowacourse.auth.domain.Nickname;
-import woowacourse.auth.domain.Password;
 import woowacourse.auth.dto.request.LoginRequest;
 import woowacourse.auth.dto.request.MemberCreateRequest;
-import woowacourse.auth.dto.request.MemberUpdateRequest;
 import woowacourse.auth.dto.request.PasswordCheckRequest;
-import woowacourse.auth.dto.request.PasswordUpdateRequest;
-import woowacourse.auth.dto.response.PasswordCheckResponse;
 import woowacourse.auth.dto.response.LoginResponse;
-import woowacourse.auth.dto.response.MemberResponse;
-import woowacourse.auth.exception.AuthorizationException;
+import woowacourse.auth.dto.response.PasswordCheckResponse;
 import woowacourse.auth.support.JwtTokenProvider;
 
 @Service
@@ -51,15 +45,11 @@ public class AuthService {
 
     @Transactional(readOnly = true)
     public LoginResponse login(LoginRequest loginRequest) {
-        Member member = findByEmail(loginRequest.getEmail(), new IllegalArgumentException("이메일과 비밀번호를 확인해주세요."));
+        Member member = memberDao.findByEmail(loginRequest.getEmail())
+                .orElseThrow(() -> new IllegalArgumentException("이메일과 비밀번호를 확인해주세요."));
         validatePassword(member.getPassword(), loginRequest.getPassword());
         String token = jwtTokenProvider.createToken(member.getEmail());
         return new LoginResponse(token, member.getNickname());
-    }
-
-    private Member findByEmail(String email, RuntimeException exception) {
-        return memberDao.findByEmail(email)
-                .orElseThrow(() -> exception);
     }
 
     private void validatePassword(String input, String saved) {
@@ -75,38 +65,5 @@ public class AuthService {
         boolean result = member.getPassword()
                 .equals(passwordCheckRequest.getPassword());
         return new PasswordCheckResponse(result);
-    }
-
-    public MemberResponse findMember(String email) {
-        Member member = findByEmail(email, new AuthorizationException("회원 정보를 찾지 못했습니다."));
-        return new MemberResponse(member);
-    }
-
-    @Transactional
-    public void updateMember(String email, MemberUpdateRequest memberUpdateRequest) {
-        validateExists(email);
-        String nickname = new Nickname(memberUpdateRequest.getNickname())
-                .getValue();
-        memberDao.updateNicknameByEmail(email, nickname);
-    }
-
-    private void validateExists(String email) {
-        if (!existsEmail(email)) {
-            throw new AuthorizationException("회원 정보를 찾지 못했습니다.");
-        }
-    }
-
-    @Transactional
-    public void updatePassword(String email, PasswordUpdateRequest passwordUpdateRequest) {
-        validateExists(email);
-        String password = new Password(passwordUpdateRequest.getPassword())
-                .getValue();
-        memberDao.updatePasswordByEmail(email, password);
-    }
-
-    @Transactional
-    public void delete(String email) {
-        validateExists(email);
-        memberDao.deleteByEmail(email);
     }
 }
