+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -3,9 +3,7 @@ package woowacourse.auth.ui;
 import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,13 +12,10 @@ import org.springframework.web.bind.annotation.RestController;
 import woowacourse.auth.application.AuthService;
 import woowacourse.auth.dto.request.LoginRequest;
 import woowacourse.auth.dto.request.MemberCreateRequest;
-import woowacourse.auth.dto.request.MemberUpdateRequest;
 import woowacourse.auth.dto.request.PasswordCheckRequest;
-import woowacourse.auth.dto.request.PasswordUpdateRequest;
 import woowacourse.auth.dto.response.EmailUniqueCheckResponse;
-import woowacourse.auth.dto.response.PasswordCheckResponse;
 import woowacourse.auth.dto.response.LoginResponse;
-import woowacourse.auth.dto.response.MemberResponse;
+import woowacourse.auth.dto.response.PasswordCheckResponse;
 import woowacourse.auth.support.AuthenticationPrincipal;
 
 @RestController
@@ -55,33 +50,5 @@ public class AuthController {
     public ResponseEntity<PasswordCheckResponse> confirmPassword(@AuthenticationPrincipal String payload,
                                                                  @RequestBody PasswordCheckRequest passwordCheckRequest) {
         return ResponseEntity.ok(authService.checkPassword(payload, passwordCheckRequest));
-    }
-
-    @GetMapping("/members/me")
-    public ResponseEntity<MemberResponse> showMember(@AuthenticationPrincipal String payload) {
-        return ResponseEntity.ok(authService.findMember(payload));
-    }
-
-    @PatchMapping("/members/me")
-    public ResponseEntity<Void> updateMember(@AuthenticationPrincipal String payload,
-                                             @RequestBody @Valid MemberUpdateRequest memberUpdateRequest) {
-        authService.updateMember(payload, memberUpdateRequest);
-        return ResponseEntity.noContent()
-                .build();
-    }
-
-    @PatchMapping("/members/password")
-    public ResponseEntity<Void> updatePassword(@AuthenticationPrincipal String payload,
-                                               @RequestBody @Valid PasswordUpdateRequest passwordUpdateRequest) {
-        authService.updatePassword(payload, passwordUpdateRequest);
-        return ResponseEntity.noContent()
-                .build();
-    }
-
-    @DeleteMapping("/members/me")
-    public ResponseEntity<Void> deleteMember(@AuthenticationPrincipal String payload) {
-        authService.delete(payload);
-        return ResponseEntity.noContent()
-                .build();
     }
 }
