+++ src/test/java/woowacourse/auth/application/AuthServiceTest.java
@@ -1,7 +1,6 @@
 package woowacourse.auth.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
@@ -13,12 +12,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.dto.request.LoginRequest;
 import woowacourse.auth.dto.request.MemberCreateRequest;
-import woowacourse.auth.dto.request.MemberUpdateRequest;
 import woowacourse.auth.dto.request.PasswordCheckRequest;
-import woowacourse.auth.dto.request.PasswordUpdateRequest;
 import woowacourse.auth.dto.response.LoginResponse;
-import woowacourse.auth.dto.response.MemberResponse;
-import woowacourse.auth.exception.AuthorizationException;
 
 @SpringBootTest
 @Transactional
@@ -47,7 +42,7 @@ class AuthServiceTest {
                 .hasMessage("이미 존재하는 이메일 주소입니다.");
     }
 
-    @DisplayName("로그인에 성공하면 jwt 토큰과 닉네임을 반환한다.")
+    @DisplayName("로그인에 성공하면 토큰과 닉네임을 반환한다.")
     @Test
     void login() {
         MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
@@ -73,7 +68,7 @@ class AuthServiceTest {
                 .hasMessage("이메일과 비밀번호를 확인해주세요.");
     }
 
-    @DisplayName("토큰과 비밀번호를 받아서, 올바른지 반환한다.")
+    @DisplayName("이메일과 비밀번호를 받아서, 올바른지 반환한다.")
     @ParameterizedTest
     @CsvSource({"1q2w3e4r!, true", "asda1234!, false"})
     void checkPassword(String password, boolean expected) {
@@ -85,133 +80,5 @@ class AuthServiceTest {
                 .isSuccess();
 
         assertThat(actual).isEqualTo(expected);
-    }
-
-
-    @DisplayName("이메일과 수정할 회원 정보를 받아 회원 정보를 수정한다.")
-    @Test
-    void updateMember() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        authService.updateMember("abc@woowahan.com", new MemberUpdateRequest("바뀐닉네임"));
-        LoginRequest loginRequest = new LoginRequest("abc@woowahan.com", "1q2w3e4r!");
-        String nickname = authService.login(loginRequest)
-                .getNickname();
-
-        assertThat(nickname).isEqualTo("바뀐닉네임");
-    }
-
-    @DisplayName("존재하지 않는 회원의 정보를 수정하려고 하면 예외를 반환한다.")
-    @Test
-    void updateMember_NotFoundMember() {
-        assertThatThrownBy(() -> authService.updateMember("abc@woowahan.com", new MemberUpdateRequest("바뀐닉네임")))
-                .isInstanceOf(AuthorizationException.class)
-                .hasMessage("유효하지 않은 토큰입니다.");
-    }
-
-    @DisplayName("올바르지 않은 형식의 닉네임으로 변경하려고 하면 예외를 반환한다.")
-    @Test
-    void updatePassword_InvalidNicknameFormat() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        assertThatThrownBy(() -> authService.updateMember("abc@woowahan.com", new MemberUpdateRequest("1234")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("닉네임 형식이 올바르지 않습니다.");
-    }
-
-    @DisplayName("이메일과 비밀번호를 받아 비밀번호를 수정한다.")
-    @Test
-    void updatePassword() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        authService.updatePassword("abc@woowahan.com", new PasswordUpdateRequest("1q2w3e4r@"));
-
-        LoginRequest loginRequest = new LoginRequest("abc@woowahan.com", "1q2w3e4r@");
-
-        assertThatCode(() -> authService.login(loginRequest))
-                .doesNotThrowAnyException();
-    }
-
-    @DisplayName("회원의 정보를 반환한다.")
-    @Test
-    void findMember() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        MemberResponse memberResponse = authService.findMember("abc@woowahan.com");
-
-        assertThat(memberResponse.getEmail()).isEqualTo("abc@woowahan.com");
-        assertThat(memberResponse.getNickname()).isEqualTo("닉네임");
-    }
-
-    @DisplayName("존재하지 않는 회원의 정보를 요청하면 예외를 반환한다.")
-    @Test
-    void findMember_NotFoundMember() {
-        assertThatThrownBy(() -> authService.findMember("abc@woowahan.com"))
-                .isInstanceOf(AuthorizationException.class)
-                .hasMessage("유효하지 않은 토큰입니다.");
-    }
-
-    @DisplayName("존재하지 않는 회원의 비밀번호를 수정하려고 하면 예외를 반환한다.")
-    @Test
-    void updatePassword_NotFoundMember() {
-        assertThatThrownBy(() -> authService.updatePassword("abc@woowahan.com", new PasswordUpdateRequest("1q2w3e4r@")))
-                .isInstanceOf(AuthorizationException.class)
-                .hasMessage("유효하지 않은 토큰입니다.");
-    }
-
-    @DisplayName("올바르지 않은 형식의 비밀번호로 변경하려고 하면 예외를 반환한다.")
-    @Test
-    void updatePassword_InvalidPasswordFormat() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        assertThatThrownBy(() -> authService.updatePassword("abc@woowahan.com", new PasswordUpdateRequest("1234")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("비밀번호 형식이 올바르지 않습니다.");
-    }
-
-    @DisplayName("이메일이 일치하는 회원 정보를 삭제한다.")
-    @Test
-    void deleteMember() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        authService.delete("abc@woowahan.com");
-        boolean actual = authService.existsEmail("abc@woowahan.com");
-
-        assertThat(actual).isFalse();
-    }
-
-    @DisplayName("존재하는 이메일인지 반환한다.")
-    @ParameterizedTest
-    @CsvSource({"abc@woowahan.com, true", "abc@naver.com, false"})
-    void existsEmail(String email, boolean expected) {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
-
-        boolean actual = authService.existsEmail(email);
-
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    @DisplayName("올바르지 않은 이메일 형식으로 이메일이 존재하는지 확인하려하면 예외를 반환한다.")
-    @Test
-    void existsEmail_InvalidFormat() {
-        String invalid = "abc";
-        assertThatThrownBy(() -> authService.existsEmail(invalid))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이메일 형식이 올바르지 않습니다.");
-    }
-
-    @DisplayName("존재하지 않는 회원을 삭제하려 하면 예외를 반환한다.")
-    @Test
-    void deleteMember_NotFoundMember() {
-        assertThatThrownBy(() -> authService.delete("abc@woowahan.com"))
-                .isInstanceOf(AuthorizationException.class)
-                .hasMessage("유효하지 않은 토큰입니다.");
     }
 }
